requirements.txt
python-dotenv==1.0.0
requests==2.31.0
pyzeebe==3.0.4
flask==2.3.3
flask-cors==4.0.0
openai==1.3.0 